backend/requirements.txt
Flask
Flask-SQLAlchemy
Flask-CORS
PyMySQL
mysql-connector-python
PyJWT
flask-jwt-extended